refactor: Modified routes, refactored code, delete password-checker.js
