+++ src/apis/signup.js
@@ -4,11 +4,11 @@ import {
   UnauthorizedError,
   NotFoundError,
   ConflictError,
-} from "../error/error.js";
+} from "../utils/error.js";
 import User from "../models/user.schema.js";
 import { asyncHandler } from "../utils/async-handler.js";
-import { banningEmptySpaces } from "../utils/validation.js";
-import tokenService from "../jwt/index.js";
+import { hasWhiteSpace } from "../utils/validation.js";
+import jwt from "../utils/jwt.js";
 import bcrypt from "bcryptjs";
 
 export const signUpRouter = express.Router();
@@ -18,26 +18,15 @@ signUpRouter.post(
   asyncHandler(async (req, res) => {
     const { email, nickname, password } = req.body.user;
 
-    if (banningEmptySpaces(email) || banningEmptySpaces(nickname) || banningEmptySpaces(password)) {
+    if (hasWhiteSpace(email) || hasWhiteSpace(nickname) || hasWhiteSpace(password)) {
       throw new ValidationError("정보 입력란에 공백을 사용할 수 없습니다.");
     }
 
-    const existingUserPromise = User.findOne({ email });
-    /**
-     * await의 역할은 비동기 코드를 동기적으로 작동하게 한다.
-     * await 안 붙이면 pending상태인 프로미스를 반환한다.
-     */
+    const existingUser = await User.findOne({
+      $or: [{ email: email }, { nickname: nickname }],
+    });
 
-    const existingNicknamePromise = User.findOne({ nickname });
-
-    const [existingUser, existingNickname] = await Promise.all([
-      existingUserPromise,
-      existingNicknamePromise,
-    ]);
-
-    // console.log(existingUser, existingNickname);
-
-    if (existingUser || existingNickname) {
+    if (existingUser) {
       throw new ConflictError("이미 사용중인 이메일 혹은 닉네임입니다.");
     }
 
@@ -57,11 +46,11 @@ signUpRouter.post(
 );
 
 signUpRouter.post(
-  "/login",
+  "/signIn",
   asyncHandler(async (req, res) => {
     const { email, password } = req.body;
 
-    if (banningEmptySpaces(email) || banningEmptySpaces(password)) {
+    if (hasWhiteSpace(email) || hasWhiteSpace(password)) {
       throw new ValidationError("정보 입력란에 공백을 사용할 수 없습니다.");
     }
 
@@ -70,12 +59,12 @@ signUpRouter.post(
       throw new NotFoundError("존재하지 않는 사용자입니다.");
     }
 
-    const comparingPassword = await bcrypt.compare(password, user.password);
-    if (!comparingPassword) {
+    const isValidPassword = await bcrypt.compare(password, user.password);
+    if (!isValidPassword) {
       throw new UnauthorizedError("비밀번호가 일치하지 않습니다.");
     }
 
-    const token = tokenService.getToken(email);
+    const token = jwt.getToken(email);
 
     res
       .set("Access-Control-Expose-Headers", "Authorization")
