+++ src/main.js
@@ -1,10 +1,11 @@
 import express from "express";
 import mongoose from "mongoose";
+import cors from "cors";
 import { companiesRouter } from "./apis/companies.js";
-import { compareRouter } from "./apis/compare.js";
 import { investmentsRouter } from "./apis/investments.js";
 import { signUpRouter } from "./apis/signup.js";
-import cors from "cors";
+import { compareRouter } from "./apis/compare.js";
+import { errorHandler } from "./utils/error-handler.js";
 
 const app = express();
 app.use(cors());
@@ -14,8 +15,6 @@ mongoose
   .then(() => console.log("Connected to DB"))
   .catch((e) => console.log(e)); //.env 연결 및 실패
 
-mongoose.connection.on("connected", () => console.log("connected"));
-mongoose.connection.on("open", () => console.log("open"));
 mongoose.connection.on("disconnected", () => console.log("disconnected"));
 mongoose.connection.on("reconnected", () => console.log("reconnected"));
 mongoose.connection.on("disconnecting", () => console.log("disconnecting"));
@@ -27,3 +26,5 @@ app.use("/investments", investmentsRouter);
 app.use("/users", signUpRouter);
 app.use("/compare", compareRouter);
 app.listen(4000, () => console.log("Server Started"));
+
+app.use(errorHandler);
