+++ src/middlewares/login-checker.js
@@ -1,5 +1,5 @@
-import { UnauthorizedError, NotFoundError } from "../error/error.js";
-import tokenService from "../jwt/index.js";
+import { UnauthorizedError, NotFoundError } from "../utils/error.js";
+import jwt from "../utils/jwt.js";
 import User from "../models/user.schema.js";
 
 export const loginChecker = async (req, res, next) => {
@@ -16,14 +16,13 @@ export const loginChecker = async (req, res, next) => {
   }
 
   try {
-    const { email, nickname } = tokenService.getPayload(token);
+    const { email, nickname } = jwt.getPayload(token);
     const user = await User.findOne({ email }).exec();
     if (!user) {
       throw new NotFoundError("사용자를 찾을 수 없습니다.");
     }
 
-    req.email = email;
-    req.nickname = nickname;
+    req.user = { email, nickname };
 
     next();
   } catch (err) {
